mod : 미션을 수행하면 화면이 넘어가게 수정
나는 GOAT야
