--- lib/dummy/dummy_data.dart
@@ -1,6 +1,0 @@
-List<String> texts = ["첫번째", "두번째", "세번째", "네번째"];
-List<String> imgs = [
-  "assets/image/img_1.png",
-  "assets/image/img_2.png",
-  "assets/image/img_1.png",
-];
