+++ lib/pages/showFairytale.dart
@@ -1,9 +1,9 @@
 import 'dart:typed_data';
 import 'package:flutter/material.dart';
 import 'package:frontend/pages/makingFairytale.dart';
-import 'package:frontend/mediapipe/pose_detector_view.dart';
+import 'package:frontend/services/mediapipe/pose_detector_view.dart';
 import '../models/scene_model.dart';
-import "package:frontend/dummy/dummy_data.dart";
+import 'package:frontend/constants/dummy_data.dart';
 
 late final SceneModel gSceneModel;
 
@@ -56,96 +56,89 @@ class Story extends StatefulWidget {
 }
 
 class _StoryState extends State<Story> {
-  int index = 0;
-  int max = 3;
+  //int index = 0;
+  int max = 8;
 
   @override
   Widget build(BuildContext context) {
     var images = widget.images;
 
-    Future.delayed(
-      const Duration(milliseconds: 10000),
-      () {
-        if (index < max - 1) {
-          setState(
-            () {
-              index += 1;
-            },
-          );
-        }
-      },
-    );
-
-    return Container(
-      child: Row(
-        mainAxisAlignment: MainAxisAlignment.center,
-        children: [
-          Flexible(
-            child: Container(
-              color: Colors.white,
-              child: Stack(
-                children: [
-                  Positioned(
-                    left: locX.elementAt(index) - 100,
-                    right: locY.elementAt(index) - 100,
-                    child: Container(
-                      width: 500,
-                      height: 500,
-                      child: Image.asset(
-                        imgs.elementAt(index),
-                      ),
-//                      child: Image.asset(gSceneModel.b64_images[index]),
-                    ),
-                  ),
-                  Positioned(
-                    left: 0,
-                    top: 0,
-                    child: Container(
-                      height: 1000,
-                      width: 1000,
-                      child: Transform.scale(
-                        scale: 0.8,
-                        child: Transform.rotate(
-                          angle: 3.141592 * (3 / 2),
-                          child: PoseDetectorView(images: images),
+    ValueNotifier(clr_index);
+    return ValueListenableBuilder<int>(
+      valueListenable: clr_index,
+      builder: (context, value, _) {
+        return Container(
+          child: Row(
+            mainAxisAlignment: MainAxisAlignment.center,
+            children: [
+              Flexible(
+                child: Container(
+                  color: Colors.white,
+                  child: Stack(
+                    children: [
+                      Positioned(
+                        left: locX.elementAt(value) - 100,
+                        right: locY.elementAt(value) - 100,
+                        child: Container(
+                          width: 500,
+                          height: 500,
+                          child: Image.asset(
+                            imgs.elementAt(value),
+                          ),
+                          //                      child: Image.asset(gSceneModel.b64_images[index]),
                         ),
                       ),
-                    ),
-                  ),
-                  Positioned(
-                    bottom: 0,
-                    left: 0,
-                    child: Camera(),
-                  )
-                ],
-              ),
-            ),
-          ),
-          Flexible(
-            child: Container(
-              width: double.infinity,
-              height: double.infinity,
-              child: Column(
-                children: [
-                  Flexible(
-                    flex: 3,
-                    child: Container(
-                      height: double.infinity,
-                      width: double.infinity,
-                      alignment: Alignment.center,
-                      child: Text(
-                        texts.elementAt(index),
-//                        gSceneModel.scriptModelList[index].scene_contents,
-                        style: TextStyle(fontSize: 40),
+                      Positioned(
+                        left: 0,
+                        top: 0,
+                        child: Container(
+                          height: 1000,
+                          width: 1000,
+                          child: Transform.scale(
+                            scale: 0.8,
+                            child: Transform.rotate(
+                              angle: 3.141592 * (3 / 2),
+                              child: PoseDetectorView(images: images),
+                            ),
+                          ),
+                        ),
                       ),
-                    ),
+                      Positioned(
+                        bottom: 0,
+                        left: 0,
+                        child: Camera(),
+                      )
+                    ],
                   ),
-                ],
+                ),
               ),
-            ),
+              Flexible(
+                child: Container(
+                  width: double.infinity,
+                  height: double.infinity,
+                  child: Column(
+                    children: [
+                      Flexible(
+                        flex: 3,
+                        child: Container(
+                          height: double.infinity,
+                          width: double.infinity,
+                          alignment: Alignment.center,
+                          child: Text(
+                            texts.elementAt(value),
+                            //                        gSceneModel.scriptModelList[index].scene_contents,
+                            style: TextStyle(fontSize: 40),
+                          ),
+                        ),
+                      ),
+                    ],
+                  ),
+                ),
+              ),
+            ],
           ),
-        ],
-      ),
+        );
+      },
     );
   }
 }
