+++ lib/services/mediapipe/camera_view.dart
@@ -1,0 +1,232 @@
+import 'dart:io';
+
+import 'package:camera/camera.dart';
+import 'package:flutter/foundation.dart';
+import 'package:flutter/material.dart';
+import 'package:google_mlkit_commons/google_mlkit_commons.dart';
+
+import 'package:frontend/main.dart';
+
+CameraController? controller;
+
+// 카메라 화면
+class CameraView extends StatefulWidget {
+  const CameraView(
+      {Key? key,
+      required this.customPaint,
+      required this.onImage,
+      this.initialDirection = CameraLensDirection.front})
+      : super(key: key);
+
+  // 스켈레톤을 그려주는 객체
+  final CustomPaint? customPaint;
+
+  // 이미지 받을 때마다 실행하는 함수
+  final Function(InputImage inputImage) onImage;
+
+  // 카메라 렌즈 방향 변수
+  final CameraLensDirection initialDirection;
+
+  @override
+  State<CameraView> createState() => _CameraViewState();
+}
+
+class _CameraViewState extends State<CameraView> {
+  // 카메라를 다루기 위한 변수
+
+  // 카메라 인덱스
+  int _cameraIndex = -1;
+
+  // 확대 축소 레벨
+  double zoomLevel = 0.0, minZoomLevel = 0.0, maxZoomLevel = 0.0;
+
+  // 카메라 렌즈 변경 변수
+  bool _changingCameraLens = false;
+
+  @override
+  void initState() {
+    super.initState();
+
+    // 카메라 설정. 기기에서 실행 가능한 카메라, 카메라 방향 설정...
+    if (cameras.any(
+      (element) =>
+          element.lensDirection == widget.initialDirection &&
+          element.sensorOrientation == 90,
+    )) {
+      _cameraIndex = cameras.indexOf(
+        cameras.firstWhere((element) =>
+            element.lensDirection == widget.initialDirection &&
+            element.sensorOrientation == 90),
+      );
+    } else {
+      for (var i = 0; i < cameras.length; i++) {
+        if (cameras[i].lensDirection == widget.initialDirection) {
+          _cameraIndex = i;
+          break;
+        }
+      }
+    }
+
+    // 카메라 실행 가능하면 포즈 추출 시작
+    if (_cameraIndex != -1) {
+      _startLiveFeed();
+    }
+  }
+
+  @override
+  void dispose() {
+    _stopLiveFeed();
+    super.dispose();
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return Container(alignment: Alignment.bottomLeft, child: _liveFeedBody());
+  }
+
+  // 전면<->후면 카메라 변경 버튼
+  Widget? _floatingActionButton() {
+    if (cameras.length == 1) return null;
+    return SizedBox(
+        height: 30.0,
+        width: 30.0,
+        child: FloatingActionButton(
+          onPressed: _switchLiveCamera,
+          child: Icon(
+            Platform.isIOS
+                ? Icons.flip_camera_ios_outlined
+                : Icons.flip_camera_android_outlined,
+            size: 20,
+          ),
+        ));
+  }
+
+  // 카메라 화면 보여주기 + 화면에서 실시간으로 포즈 추출
+  Widget _liveFeedBody() {
+    if (controller?.value.isInitialized == false) {
+      return Container();
+    }
+
+    final size = MediaQuery.of(context).size;
+    // 화면 및 카메라 비율에 따른 스케일 계산
+    // 원문: calculate scale depending on screen and camera ratios
+    // this is actually size.aspectRatio / (1 / camera.aspectRatio)
+    // because camera preview size is received as landscape
+    // but we're calculating for portrait orientation
+    var scale = size.aspectRatio * controller!.value.aspectRatio;
+
+    // to prevent scaling down, invert the value
+    if (scale < 1) scale = 1 / scale;
+
+    return Container(
+      child: Stack(
+        fit: StackFit.expand,
+        children: <Widget>[
+          // 전면 후면 변경 시 화면 변경 처리
+          Transform.scale(
+            scale: scale,
+            child: Center(
+              child: _changingCameraLens
+                  ? const Center(
+                      child: Text('Changing camera lens'),
+                    )
+                  : Container(),
+            ),
+          ),
+          // 추출된 스켈레톤 그리기
+          if (widget.customPaint != null) widget.customPaint!,
+
+          // if (_byteImage != null)
+          //   Image.memory(_byteImage!, width: 100, height: 100),
+          // if (_byteImage != null)
+          //   Image(image: MemoryImage(_byteImage!), width: 100, height: 100),
+        ],
+      ),
+    );
+  }
+
+  // 실시간으로 카메라에서 이미지 받기(비동기적)
+  Future _startLiveFeed() async {
+    final camera = cameras[_cameraIndex];
+    controller = CameraController(
+      camera,
+      ResolutionPreset.high,
+      enableAudio: false,
+    );
+    controller?.initialize().then((_) {
+      if (!mounted) {
+        return;
+      }
+      controller?.getMinZoomLevel().then((value) {
+        zoomLevel = value;
+        minZoomLevel = value;
+      });
+      controller?.getMaxZoomLevel().then((value) {
+        maxZoomLevel = value;
+      });
+      // 이미지 받은 것을 _processCameraImage 함수로 처리
+      controller?.startImageStream(_processCameraImage);
+      setState(() {});
+    });
+  }
+
+  Future _stopLiveFeed() async {
+    await controller?.stopImageStream();
+    await controller?.dispose();
+    controller = null;
+  }
+
+  // 전면<->후면 카메라 변경 함수
+  Future _switchLiveCamera() async {
+    setState(() => _changingCameraLens = true);
+    _cameraIndex = (_cameraIndex + 1) % cameras.length;
+
+    await _stopLiveFeed();
+    await _startLiveFeed();
+    setState(() => _changingCameraLens = false);
+  }
+
+  // 카메라에서 실시간으로 받아온 이미치 처리: PoseDetectorView에서 받아온 함수인 onImage(이미지에 포즈가 추출되었으면 스켈레톤 그려주는 함수) 실행
+  Future _processCameraImage(CameraImage image) async {
+    final WriteBuffer allBytes = WriteBuffer();
+    for (final Plane plane in image.planes) {
+      allBytes.putUint8List(plane.bytes);
+    }
+    final bytes = allBytes.done().buffer.asUint8List();
+
+    final Size imageSize =
+        Size(image.width.toDouble(), image.height.toDouble());
+
+    final camera = cameras[_cameraIndex];
+    final imageRotation =
+        InputImageRotationValue.fromRawValue(camera.sensorOrientation);
+    if (imageRotation == null) return;
+
+    final inputImageFormat =
+        InputImageFormatValue.fromRawValue(image.format.raw);
+    if (inputImageFormat == null) return;
+
+    final planeData = image.planes.map(
+      (Plane plane) {
+        return InputImagePlaneMetadata(
+          bytesPerRow: plane.bytesPerRow,
+          height: plane.height,
+          width: plane.width,
+        );
+      },
+    ).toList();
+
+    final inputImageData = InputImageData(
+      size: imageSize,
+      imageRotation: imageRotation,
+      inputImageFormat: inputImageFormat,
+      planeData: planeData,
+    );
+
+    final inputImage =
+        InputImage.fromBytes(bytes: bytes, inputImageData: inputImageData);
+
+    // PoseDetectorView에서 받아온 함수인 onImage(이미지에 포즈가 추출되었으면 스켈레톤 그려주는 함수) 실행
+    widget.onImage(inputImage);
+  }
+}
